Rename add-trip.js to add-trip-modal.js.
