+++ frontend/src/components/ViewTrips/add-trip-modal.js
